# Perf/dict value