Addresses #18.

The `dict.Value` now performs a binary search on the underlying key-value structure instead of iterating linearly over all entries. This reduces the complexity from O(n) to O(log n) and brings a significant performance improvement when dictionaries grow large.

On a 1024-entry dictionary, typical lookups dropped from multiple microseconds to <100 ns (36–61× faster on mid/edge hits and misses).

```text
# before

BenchmarkDictValue/size_1024_hit_first          137312391                9.115 ns/op           0 B/op          0 allocs/op
BenchmarkDictValue/size_1024_hit_mid              385447              2709 ns/op               0 B/op          0 allocs/op
BenchmarkDictValue/size_1024_hit_last             225544              5386 ns/op               0 B/op          0 allocs/op
BenchmarkDictValue/size_1024_miss_low             463954              2585 ns/op               0 B/op          0 allocs/op
BenchmarkDictValue/size_1024_miss_high            225627              5474 ns/op               0 B/op          0 allocs/op

# after

BenchmarkDictValue/size_1024_hit_first          15694873                78.60 ns/op            0 B/op          0 allocs/op
BenchmarkDictValue/size_1024_hit_mid            16182238                74.46 ns/op            0 B/op          0 allocs/op
BenchmarkDictValue/size_1024_hit_last           14287293                83.86 ns/op            0 B/op          0 allocs/op
BenchmarkDictValue/size_1024_miss_low           17558726                68.35 ns/op            0 B/op          0 allocs/op
BenchmarkDictValue/size_1024_miss_high          13268202                89.83 ns/op            0 B/op          0 allocs/op
```